Add BinaryColumnBase
